Template method pattern in Pea, FreezePea
