+++ src/FreezePea.java
@@ -3,26 +3,13 @@ import java.awt.*;
 /**
  * Created by Armin on 6/28/2016.
  */
-public class FreezePea extends Pea {
+public class FreezePea extends AbstractPea {
 
     public FreezePea(GamePanel parent, int lane, int startX) {
         super(parent, lane, startX);
     }
 
-    @Override
-    public void advance() {
-        Rectangle pRect = new Rectangle(getPosX(), 130 + getMyLane() * 120, 28, 28);
-        for (int i = 0; i < gp.getLaneZombies().get(getMyLane()).size(); i++) {
-            Zombie z = gp.getLaneZombies().get(getMyLane()).get(i);
-            Rectangle zRect = new Rectangle(z.getPosX(), 109 + getMyLane() * 120, 400, 120);
-
-            /* edited */
-            if(attackZombie(z, pRect,zRect,i)) break;
-        }
-
-        setPosX(getPosX() - 15);
+    protected void setPosXX() {
+    	setPosX(getPosX() - 15);
     }
-
-
-
 }
